Refactor DropDown
