+++ src/components/formFields/check-box.tsx
@@ -9,20 +9,22 @@ interface CheckboxProps {
 }
 
 const Checkbox = (props: CheckboxProps) => {
+    const handleClick = () => {
+        if (!props.disabled) {
+            props.onChange(!props.checked);
+        }
+    };
+
     return (
         <div
             className={
                 'w-12 h-12 p-2 transition-colors duration-150 ' +
                 (props.disabled
-                    ? 'text-gray-500 cursor-not-allowed'
+                    ? 'text-gray-500 cursor-not-allowed '
                     : 'text-gray-800 cursor-pointer ') +
                 (props.className ? props.className : '')
             }
-            onClick={() => {
-                if (!props.disabled) {
-                    props.onChange(!props.checked);
-                }
-            }}
+            onClick={handleClick}
         >
             {props.checked ? icons.checkbox_true : icons.checkbox_false}
         </div>
